requirements.txt
langchain>=0.3.0
langchain-groq>=0.2.0
langchain-core>=0.3.0
sentence-transformers>=3.0.0
faiss-cpu>=1.7.4
numpy>=1.24.0
pymupdf>=1.24.0
requests>=2.31.0
